+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ### Added
 
-- #16 - changed `MaintenanceNotification.raw` from `str` to `bytes`
+- #16 - changed `MaintenanceNotification.raw` from `str` to `bytes`, improve **Zayo** parser, and add -v/--verbose CLI option
 - #17 - added **Lumen** parser and refactor HTML parser `process` method
 - #18 - added **Telstra** parser and made `Maintenance.stamp` attribute optional
 - #19 - added **Megaport**
+
+### Fixes
+
+- #15 - Update Pydantic version due to security advisory GHSA-5jqp-qgf6-3pvh
 
 ## v1.0.2 - 2021-05-05